begin v1 rewrite
only accept .md extensions; handle pathing better; organize element varibles; update styles due to tag updates
