+++ css/index.css
@@ -20,9 +20,11 @@ body {
     justify-content: space-between;
     height: 32px;
     margin: 0;
+    overflow: hidden;
     padding: 0;
     position: fixed;
     top: 0;
+    white-space: nowrap;
     width: calc(100% - 363px);
     z-index: 1; }
     body > menu > container > icon {
@@ -36,10 +38,15 @@ body {
         color: #ff2052; }
         body > menu > container > icon[enabled]:hover {
           color: #d3002f; }
-    body > menu > container > text {
+    body > menu > path {
+      background: #8f8f8f;
       cursor: pointer;
+      flex-grow: 1;
       font-size: 16px;
+      padding: 4px 0;
       user-select: none; }
+      body > menu > path:empty {
+        padding: 0; }
   body > repo {
     display: inline-block;
     height: 100%;
@@ -58,7 +65,6 @@ body {
         border: 4px dashed #777;
         border-radius: 16px;
         box-sizing: border-box;
-        cursor: pointer;
         display: block;
         height: 75%;
         top: 50%;
@@ -66,8 +72,9 @@ body {
         margin: 4px;
         position: relative; }
         body > repo > container > upload > input {
-          height: 100%;
+          cursor: pointer;
           filter: opacity(0);
+          height: 100%;
           width: 100%; }
         body > repo > container > upload > info {
           align-items: center;
