+++ index.html
@@ -21,9 +21,7 @@
     <icon button="broken_image">broken_image</icon>
     <icon button="file_upload">file_upload</icon>
    </container>
-   <container>
-    <text></text>
-   </container>
+   <path></path>
    <container>
     <icon button="vertical_align_top">vertical_align_top</icon>
     <icon button="edit">edit</icon>
@@ -32,7 +30,7 @@
   <repo>
    <container>
     <upload>
-     <input type="file" accept=".markdown,.mdown,.mkdn,.md,.mkd,.mdwn,.mdtxt,.mdtext,.text,.rmd"/>
+     <input type="file" accept=".md"/>
      <info>
       <icon>file_upload</icon>
       <text>Drag-and-Drop or<br/>Click Here to Upload Your Markdown</text>
