+++ js/index.js
@@ -3,436 +3,86 @@
 (function(){
  'use strict';
 
- let toMarkdown;
+ const nw_gui = require('nw.gui');
+ const $refresh = $('icon[button="refresh"]');
+ const $autorenew = $('icon[button="autorenew"]');
+ const $broken_image = $('icon[button="broken_image"]');
+ const $file_upload = $('icon[button="file_upload"]');
+ const $vertical_align_top = $('icon[button="vertical_align_top"]');
+ const $edit = $('icon[button="edit"]');
+ const $path = $('path');
 
- const gui = require('nw.gui');
+ showdown.setFlavor("github");
 
- const _FR = new FileReader();
- const $container_article = $('repo > container');
- const $article = $('article');
+ const _FileReader = new FileReader();
+ const $container = $('repo > container');
  const $upload = $('upload');
  const $input = $('input');
- //const $history_button = $('icon[button="history"]');
- const $upload_button = $('icon[button="file_upload"]');
- const $refresh_button = $('icon[button="refresh"]');
- const $image_button = $('icon[button="broken_image"]');
- const $autorenew_button = $('icon[button="autorenew"]');
- //const $error_button = $('icon[button="error_outline"]');
- const $top_button = $('icon[button="vertical_align_top"]');
- const $edit_button = $('icon[button="edit"]');
- const $path = $('menu > container > text');
+ const $article = $('article');
 
- const $links_listInternal = $('#_linkInternal');
- const $links_listDead = $('#_linkDead');
- const $images_listBroken = $('#_imageBroken');
- const $images_listPng = $('#_imagePng');
- const $images_listWide = $('#_imageWide');
+ const paths = Object.seal({
+  "directory": null,// C:\\...\osu-wiki\wiki\
+  "locale": null,// en.md
+  "path": null,// ~\osu-wiki\...\en.md
+  "root": null,// ~\osu-wiki\
+  "working": null// C:\\...\osu-wiki\wiki\...\en.md
+ });
 
- let thisfile;
- let path;
- let directory;
- let working;
- let locale;
-
- let history =[];
-
- let imageSuffix = "?t=" + Math.random();
- let oldImageSuffix = imageSuffix;
-
- _FR.onload = function(){
-
-  thisfile = thisfile.replace(/\//ig, "\\").replace(/\\\\/ig, "\\");
-  history.push(thisfile);
-
-  $upload.toggle(false);
-  $article.toggle(true);
-  let _lines = this.result;
-
-  showdown.setFlavor("github");
+ _FileReader.onload = function(){
   toMarkdown = new showdown.Converter({
    "prefixHeaderId": false
   });
 
-  let _html = toMarkdown.makeHtml(_lines);
-
-  $path.empty().append(path);
-  $article.empty();
-
+  let _html = toMarkdown.makeHtml(this.result);
   $article.append(_html);
 
-  $links_listInternal.empty();
-  $links_listDead.empty();
-  $images_listBroken.empty();
-  $images_listPng.empty();
-  $images_listWide.empty();
-
-  //TODO history
-  if( isImage ){
-   oldImageSuffix = imageSuffix;
-   imageSuffix = "?t=" + Math.random();
-  }
-  for( let i = 0; i < $('img').length; i++ ){
-   let _target = $($('img')[i]);
-   let src = _target.attr("src");
-   if( /^(\.(\/|\\))/ig.test(src) ){
-    _target.prop("src", directory + "\\" + src.substring(2, src.length));
-   }
-   else if( /^[a-zA-Z0-9_]/ig.test(src) ){
-    _target.prop("src", directory + "\\" + src);
-   }
-   else if( /\\|\//g.test(src) ){
-    _target.prop("src", working + "\\" + src);
-   }
-   else{
-    _target.prop("src", directory + "\\" + src);
-   }
-   if( isImage ){
-    _target.prop("src", _target.prop("src") + imageSuffix);
-   }else{
-    _target.prop("src", _target.prop("src") + oldImageSuffix);
-   }
-  }
-
-  $('img')
-   .on("load", function(e){
-   if( e.target.naturalWidth > $article.outerWidth() ){
-    let src = $(e.target).prop("src");
-    let _li = $('<li/>').append(src.substring(src.lastIndexOf("/wiki/"), src.lastIndexOf("?t")));
-    $images_listWide.append(_li);
-   }
-   if( $(e.target).prop("src").split('.').pop() === "png" ){
-    let src = $(e.target).prop("src");
-    let _li = $('<li/>').append(src.substring(src.lastIndexOf("/wiki/"), src.lastIndexOf("?t")));
-    $images_listPng.append(_li);
-   }
-  })
-   .on("error", function(e){
-   let src = $(e.target).prop("src");
-   let _li = $('<li/>').append(src.substring(src.lastIndexOf("/wiki/"), src.lastIndexOf("?t")));
-   //let _li = $('<li/>').append($(e.target).prop("src").substring($(e.target).prop("src").lastIndexOf("?time=", 0)));
-   $images_listBroken.append(_li);
-  })
-  ;
-
-  (function links(i){
-   if( i < $('a').length ){
-    //TOFIX section links will fail this (adds \en.md when it shouldn't)
-    let _target = $($('a')[i]);
-    let href = _target.attr("href");
-    if( /^\.\/|^\.\\/ig.test(href) ){
-     href = href.substring(2, href.length);
-    }
-    if( /^http/ig.test(href) ){
-     exLinks(_target);
-    }
-    else if( /^mailto/ig.test(href) ){
-     exLinks(_target);
-    }
-    else if( /^\\|^\//g.test(href) ){
-     if( /\.(md|jpg|png|webp|bmp|pdf|mp3|ogg|wav|mp4|webm|avi)\/?/ig.test(href) ){
-      _target.prop("href", directory + "\\" + _target.attr("href"));
-      exLinks(_target);
-     }else{
-      _target.prop("href", working + "\\" + _target.attr("href") + locale);
-      inLinks(_target, true);
-     }
-    }
-    else if( /^\.(\\|\/)/ig ){
-     if( /\.(md|jpg|png|webp|bmp|pdf|mp3|ogg|wav|mp4|webm|avi)\/?/ig.test(href) ){
-      let _li = $('<li/>').append(href);
-      $links_listInternal.append(_li);
-      _target.prop("href", directory + "\\" + _target.attr("href"));
-      exLinks(_target);
-     }else{
-      _target.prop("href", directory + "\\" + _target.attr("href").substring(_target.attr("href").lastIndexOf("./"), _target.attr("href").length) + locale);
-      inLinks(_target, false);
-     }
-    }
-    else{
-     if( /\.(md|jpg|png|webp|bmp|pdf|mp3|ogg|wav|mp4|webm|avi)\/?/ig.test(href) ){
-      let _li = $('<li/>').append(href);
-      $links_listInternal.append(_li);
-      _target.prop("href", directory + "\\" + _target.attr("href"));
-      exLinks(_target);
-     }else{
-      _target.prop("href", directory + "\\" + _target.attr("href") + locale);
-      inLinks(_target, false);
-     }
-    }
-
-    testLink(_target.attr("href"));
-
-    i++;
-    window.setTimeout(function(){
-     links(i);
-    }, 100);
-   }
-  })(0);
  };
 
- function testLink(target){
-  // ignore *.ppy.sh links (they are too common)
-  if( !/ppy\.sh/ig.test(target) ){
-   $.ajax({
-    "type": "HEAD",
-    "url": target,
-    "error":function(){
-     let _li = $('<li/>').append(target.replace(/\//ig, "\\").replace(/\\\\/ig, "\\"));
-     $links_listDead.append(_li);
-    }
-   });
-  }
- }
+ (function(){
+  $input.on('change', function(){
+   if( this.files.length > 0 ){
+    let file = this.files[0];
+    let file_path = file.path;
+    if( /\.md$/.test(file.name) ){
+     // C:\\...\osu-wiki\wiki\
+     paths.directory = file_path.substring(0, Number(file_path.lastIndexOf("\\")) + 1);
+     // en.md
+     paths.locale = file_path.substring(Number(file_path.lastIndexOf("\\")) + 1, file_path.length);
+     // \osu-wiki\wiki\...\en.md
+     paths.path = "~" + file_path.substring(file_path.indexOf("\\osu-wiki\\"), file_path.length);
+     // ~\osu-wiki\
+     paths.root = file_path.substring(0, file_path.lastIndexOf("\\wiki\\"));
+     // C:\\...\osu-wiki\wiki\...\
+     paths.working = file_path;
 
- function exLinks(target){
-  target.on("click", function(e){
-   e.preventDefault();
+     console.log(paths);
+     
+     $path.empty().append(paths.path);
+    }
+
+   }
   });
- }
- function inLinks(target, isRoot){
-  if( isRoot ){
-   target.on("click", function(e){
-    e.preventDefault();
-    let xhr = new XMLHttpRequest();
-    xhr.onreadystatechange = function(){
-     if( this.readyState === 4 && this.status === 0 ){
-      thisfile = target.attr("href");
-      path = "~" + target.attr("href").substring(target.attr("href").lastIndexOf("\\osu-wiki"), target.attr("href").length);
-      directory = target.attr("href").substring(0, target.attr("href").lastIndexOf("\\"));
-      locale = target.attr("href").substring(target.attr("href").lastIndexOf("\\"), target.attr("href").length);
-      _FR.readAsText(this.response);
-     }
-    };
-    xhr.open('GET', target.attr("href"));
-    xhr.responseType = 'blob';
-    xhr.send();
-   });
-  }else{
-   target.on("click", function(e){
-    e.preventDefault();
-    let xhr = new XMLHttpRequest();
-    xhr.onreadystatechange = function(){
-     if( this.readyState === 4 && this.status === 0 ){
-      thisfile = target.attr("href");
-      path = "~" + target.attr("href").substring(target.attr("href").lastIndexOf("\\osu-wiki"), target.attr("href").length);
-      directory = target.attr("href").substring(0, target.attr("href").lastIndexOf("\\"));
-      locale = target.attr("href").substring(target.attr("href").lastIndexOf("\\"), target.attr("href").length);
-      _FR.readAsText(this.response);
-     }
-    };
-    xhr.open('GET', target.attr("href"));
-    xhr.responseType = 'blob';
-    xhr.send();
-   });
-  }
- }
+  $refresh.on('click', function(){
 
- $input.on("change", function(e){
-  let _target = e.target;
-  let _files = _target.files;
-  if( _files.length > 0 ){
-   //TOFIX this file check sometimes doesn't do anything
-   //console.log(_files[0].name.split('.').pop(), /markdown|mdown|mkdn|md|mkd|mdwn|mdtxt|mdtext|text|rmd/ig.test(_files[0].name.split('.').pop()));
-   if( /markdown|mdown|mkdn|md|mkd|mdwn|mdtxt|mdtext|text|rmd/ig.test(_files[0].name.split('.').pop()) ){
-    // directory = from folder
-    // working = from root
-    thisfile = _files[0].path;
-    path = "~" + _files[0].path.substring(_files[0].path.lastIndexOf("\\osu-wiki"), _files[0].path.length);
-    directory = _files[0].path.substring(0, _files[0].path.lastIndexOf("\\"));
-    working = _files[0].path.substring(0, _files[0].path.lastIndexOf("\\wiki"));
-    locale = _files[0].path.substring(_files[0].path.lastIndexOf("\\"), _files[0].path.length);
+  });
+  $autorenew.on('click', function(){
 
-    _FR.readAsText(_files[0]);
-   }
-  }
- });
+  });
+  $broken_image.on('click', function(){
 
- $upload_button.on('click', function(){
-  $input.click();
- });
+  });
+  $file_upload.on('click', function(){
+   $input.trigger("click");
+  });
+  $path.on('click', function(){
 
- $refresh_button.on('click', function(){
-  if( thisfile !== undefined ){
-   let xhr = new XMLHttpRequest();
-   xhr.onreadystatechange = function(){
-    if( this.readyState === 4 && this.status === 0 ){
-     _FR.readAsText(this.response);
-    }
-   };
-   xhr.open('GET', thisfile);
-   xhr.responseType = 'blob';
-   xhr.send();
-  }
- });
+  });
+  $vertical_align_top.on('click', function(){
+   $container.scrollTop(0);
+  });
+  $edit.on('click', function(){
 
- let isImage = false;
- $image_button.on("click", function(){
-  if( thisfile !== undefined ){
-   if( isImage ){
-    $image_button.removeAttr("enabled");
-    isImage = false;
-   }else{
-    $image_button.attr("enabled", "");
-    isImage = true;
-   }
-  }
- });
+  });
 
- let isAutorenew = false, timerAutorenew;
- $autorenew_button.on("click", function(){
-  if( thisfile !== undefined ){
-   if( isAutorenew ){
-    $autorenew_button.removeAttr("enabled");
-    isAutorenew = false;
-    window.clearInterval(timerAutorenew);
-   }else{
-    $autorenew_button.attr("enabled", "");
-    isAutorenew = true;
-    timerAutorenew = window.setInterval(function(){
-     $refresh_button.trigger("click");
-    }, 1000);
-   }
-  }
- });
-
- window.ondragover = function(e) { e.preventDefault(); return false; };
- window.ondrop = function(e) {
-  e.preventDefault();
-
-  let file = e.dataTransfer.files[0];
-
-  thisfile = file.path;
-  path = "~" + file.path.substring(file.path.lastIndexOf("\\osu-wiki"), file.path.length);
-  directory = file.path.substring(0, file.path.lastIndexOf("\\"));
-  working = file.path.substring(0, file.path.lastIndexOf("\\wiki"));
-  locale = file.path.substring(file.path.lastIndexOf("\\"), file.path.length);
-
-  _FR.readAsText(file);
-
-  return false;
- };
- $path.on("click", function(){
-  if( thisfile !== undefined ){
-   gui.Shell.showItemInFolder(thisfile);
-  }
- });
-
- /*let isError = false;
- $error_button.on("click", function(){
-  if( isError ){
-   $error_button.removeAttr("enabled");
-   //TODO send signal to close window
-   isError = false;
-  }else{
-   $error_button.attr("enabled", "");
-   gui.Window.open("errors.html", {
-    "width": 400,
-    "height": 400,
-    "min_width": 400,
-    "max_width": 400,
-    //"frame": false
-   });
-   isError = true;
-  }
- });
-
- let isHistory = false;
- $history_button.on("click", function(){
-  if( isHistory ){
-   $history_button.removeAttr("enabled");
-   isHistory = false;
-  }else{
-   $history_button.attr("enabled", "");
-   isHistory = true;
-  }
- });*/
-
- $top_button.on("click", function(){
-  $container_article.scrollTop(0);
- });
- $edit_button.on("click", function(){
-  if( thisfile !== undefined ){
-   gui.Shell.openItem(thisfile);
-  }
- });
-
- $links_listInternal.on("click", "li", function(e){
-  if( thisfile !== undefined ){
-   let _target = $(e.target);
-   let _link = _target.text();
-   if( /^(\.(\/|\\))/ig.test(_link) ){
-    // TOFIX verify link check from root works
-    _link = directory + "\\" + _link.substring(2, _link.length);
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    gui.Shell.openItem(_link);
-   }
-   else if( /^(http)/ig.test(_link) ){
-    // TOFIX verify link check from root works
-    gui.Shell.openExternal(_link);
-   }
-   else if( /\\|\//g.test(_link) ){
-    // TOFIX verify link check from root works
-    _link = working;
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    gui.Shell.openItem(_link);
-   }
-   else{
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    gui.Shell.openItem(_link);
-   }
-  }
- });
- $links_listDead.on("click", "li", function(e){
-  if( thisfile !== undefined ){
-   let _target = $(e.target);
-   let _link = _target.text();
-   if( /^(http)/ig.test(_link) ){
-    gui.Shell.openExternal(_link);
-   }
-   else if( /^(\.(\/|\\))/ig.test(_link) ){
-    // TOFIX verify link check from root works
-    _link = directory + "\\" + _link.substring(2, _link.length);
-    console.log('directory', _link);
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    gui.Shell.openItem(_link);
-   }
-   else if( /^(\\|\/)/g.test(_link) ){
-    _link = working;
-    console.log('working', _link);
-    // TOFIX verify link check from root works
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    gui.Shell.openItem(_link);
-   }
-   else{
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    gui.Shell.openItem(_link);
-   }
-  }
- });
- $images_listBroken.on("click", "li", function(e){
-  if( thisfile !== undefined ){
-   let _target = $(e.target);
-   let _link = _target.text();
-   if( /^(\.(\/|\\))/ig.test(_link) ){
-    // TOFIX verify link check from root works
-    _link = directory + "\\" + _link.substring(2, _link.length);
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    gui.Shell.openItem(_link);
-   }
-   else if( /^(http)/ig.test(_link) ){
-    // TOFIX verify link check from root works
-    gui.Shell.openExternal(_link);
-   }
-   else if( /^(\\|\/)/g.test(_link) ){
-    _link = working + _link;
-    // TOFIX verify link check from root works
-    gui.Shell.openItem(_link);
-   }
-   else{
-    _link = _link.substring(0, _link.lastIndexOf("\\"));
-    //_link = directory + _link;
-    gui.Shell.openItem(_link);
-   }
-  }
- });
+ })();
 })();
